Refatora views para modularizar listagens
